soc: qcom: ipa: drop argument to ipa_interrupts_init()
There is no longer any need to access the platform device pointer
passed to ipa_interrupts_init().  The only thing that's required is
the address of the platform device's device structure, and that's
available in ipa_ctx->dev.

So drop the platform_device pointer argument, and just use the
device pointer found in the IPA context.

Signed-off-by: Alex Elder <elder@linaro.org>