soc: qcom: ipa: kill ipa_context->ipa_pdev
The only place the platform device pointer in the IPA context is
used is within ipa_plat_drv_probe().  There's no need to store it in
the IPA context, so get rid of that field and just use the passed-in
platform structure pointer instead.

Rename that argument to be "platform_device".

Signed-off-by: Alex Elder <elder@linaro.org>