soc: qcom: ipa: kill ipa_context->ipa_wrapper_size
The ipa_wrapper_size field of the ipa_context structure is only used
inside ipa_plat_drv_probe().  There's no need to save this
information in the ipa_context structure, so use a local variable
and get rid of the field.

Reorder a few fields in the IPA context structure.

Signed-off-by: Alex Elder <elder@linaro.org>