soc: qcom: ipa: validate config very early
The only piece of runtime information needed to validate the
configuration is the filter bitmask, and that's computed using
static configuration data.

Move the computation of that bitmask, and the call to config_valid()
so it occurs immediately after we've gotten past the two
initializations that can cause EPROBE_DEFER re-probes.

Signed-off-by: Alex Elder <elder@linaro.org>