soc: qcom: ipa: rework logic in ipa_init_sram()
The physical address computed in ipa_init_sram() is a little
strange.  It uses ipahal_reg_n_offset() to determine an offset, but
it passes 1/4 of the smem_restricted_bytes value.  I'm pretty sure
this is done because it assumes the n_ofst value for the
IPA_SRAM_DIRECT_ACCESS_n register is 4.

I believe what this is trying to do is compute the address of the
first usable offset in smem, after skipping over all of the
"restricted" bytes at the beginning of the IPA register space.

Assuming that, a clearer way to do that is to add the "restricted
bytes" value to the *first* offset in IPA register space.

Signed-off-by: Alex Elder <elder@linaro.org>