soc: qcom: ipa: don't pass filter_count to ipahal_flt_generate_empty_img()
The filter count passed to ipahal_flt_generate_empty_img() is simply
a count of the set bits in the bitmask.  Just have the called function
compute that based on the mask, rather than passing it as an argument.

Update a bunch of comments in that function.  Also remove the
comments from the header, making the C file the authoritative place
to find the description of the function.

Signed-off-by: Alex Elder <elder@linaro.org>