soc: qcom: ipa: remove two assertions
In ipahal_read_reg_n_fields() there is an assertion that a function
pointer is non-null before dereferencing that function pointer.
Since the function is being called, it's pretty obvious that the
function pointer can't be NULL.  The assertion doesn't really add
value, so get rid of it.

Remove a similar assertion in ipahal_write_reg_n_fields(), for the
same reason.

Signed-off-by: Alex Elder <elder@linaro.org>