soc: qcom: ipa: kill idsym()
The ipahal_regs[] array is built up using a set of helper macros.
One of them, idsym(), simply prepends IPA_ to the name of the symbol
used for the array index.  This was helpful before, when symbol
names were much longer.  But it also has the down side of making
the symbols used in defining the array not searchable.

Add the "IPA_" to the names used in building the array, and get rid
of the idsym() macro.

Signed-off-by: Alex Elder <elder@linaro.org>