soc: qcom: ipa: change struct gsi_log_err
Change the definition of the gsi_log_err structure so standard types
are used where their bit positions properly align, using bitfields
only where necessary.  Use the smallest possible types that align
(in this case, 8-bit unsigned).  Note that this structure is packed,
but the new definition has the same layout as before.

Signed-off-by: Alex Elder <elder@linaro.org>