soc: qcom: ipa: change struct ipa_hw_error_event_data
Change the definition of the ipa_hw_error_event_data structure to
avoid using bitfields.  The register is read via the raw32b field,
but the only meaningful value is found in the first byte.  Rather
than using a bitfield (and naming the reserved portion of the
register), just define the error_type field as a byte at the front
of the union.  Note that this structure is packed, but the new
definition has the same layout as before.

Signed-off-by: Alex Elder <elder@linaro.org>