soc: qcom: ipa: change struct ipa_hw_cpu_cmd_completed_response_data
Change the definition of the ipa_hw_cpu_cmd_completed_response_data
structure to avoid using bitfields.  The register is read via the
raw32b field, but the only meaningful values are found in the first
two bytes.  Rather than using bitfields (and naming the reserved
portion of the register), just define the original_cmd_op and status
fields as a bytes in the first structure in the union.  Note that
this structure is packed, but the new definition has the same layout
as before.

Signed-off-by: Alex Elder <elder@linaro.org>