soc: qcom: ipa: kill EXTRACT_UC_FEATURE()
The EXTRACT_UC_FEATURE() macro is used in two places.  Both of them
are used to determine whether the 3-bit HW_FEATURE portion of a
response byte supplied by the microcontroller is in range.  The
macro shifts its argument right 5 bits; if it is provided an 8-bit
unsigned value, the range of possible results is 0..7 inclusive.
 - In ipa_uc_event_handler(), the value passed to the macro is an
   8-bit unsigned event_op.  It's used in a comparison to see if
   the result is IPA_HW_FEATURE_MAX or more.  IPA_HW_FEATURE_MAX is
   8, and the result of the test will always be false, so it's not
   needed.
- In ipa_uc_response_hdlr() the value passed to the macro is an
  8-bit response_op.  Here too the result of the macro is compared
  IPA_HW_FEATURE_MAX, and again the test will always yield false, so
  it can be removed.

There remain no uses of the EXTRACT_UC_FEATURE() macro, so it can be
removed.

Signed-off-by: Alex Elder <elder@linaro.org>