soc: qcom: ipa: kill ipa_uc_ctx->pending_cmd
When a microcontroller command completes, the AP receives an event
that is handled by ipa_uc_response_hdlr().  The response indicates
what command type caused the completion, and the handler verifies
this matches the last command it sent to the microcontroller.
If these don't match, the event is reported but otherwise ignored.

If the response did match the last command issued, the status of the
command is recorded in the microcontroller context.  But this status
is otherwise completely ignored in the IPA lite code.  We can
therefore get rid of this status field, planning to add it back
again when we do more interaction with the microcontroller.

In fact, the only microcontroller command we currently issue to the
microcontroller is IPA_CPU_2_HW_CMD_ERR_FATAL, which only occurs
in ipa_uc_panic_notifier() in the event of a system crash.  We don't
even wait for the response in that case, so there's no need to save
the status either.

So get rid of the pending_cmd and status fields of the microcontroller
context and just report the command and status in a log message when
they occur.

Signed-off-by: Alex Elder <elder@linaro.org>