soc: qcom: ipa: ipa_uc_interface_init() is only called once
The only caller of ipa_uc_interface_init() is ipa_post_init().  That
function is called only after the modem signals it has completed its
early IPA hardware initialization, and that should occur exactly once.

The only place the uc_inited flag of the microcontroller context
gets set is at the end of ipa_uc_interface_init().  Since it would
be an error for that to be called more than once, there's no need
for it to verify the uc_inited hasn't been set yet.

Signed-off-by: Alex Elder <elder@linaro.org>