soc: qcom: ipa: have ipa_uc_interface_init() return pointer
Have ipa_uc_interface_init() return the address of the microcontroller
context rather than an error code.  The only error is failure to map
the region of memory it shares with the AP, so a NULL pointer handles
that case just fine.  The pointer returned is still the address of
the context structure embedded in the IPA context, but that's going
to change shortly.

Rename the function ipa_uc_init().

Signed-off-by: Alex Elder <elder@linaro.org>