soc: qcom: ipa: kill IPA_HW_FEATURE_COMMON
Since IPA_HW_FEATURE_COMMON has value 0 we can get away with not
using the FEATURE_ENUM_VAL() macro to generate the event and error
codes, and command and response codes we need.

Add a comment about the way these codes are structured (as a 3-bit
feature and a 5-bit code), but explain that since all the codes we
need are "common" we don't bother manufacturing them from their two
components.

Signed-off-by: Alex Elder <elder@linaro.org>