soc: qcom: ipa: fix a very old bug
This commit was the place where send_uc_command_nowait() was used to
send the IPA_CPU_2_HW_CMD_ERR_FATAL command to the microcontroller.
  f7b8f12145a5 msm: ipa_uc: use send_uc_command_nowait() in ipa3_uc_panic_notifier()

It was also the place where a bug was introduced, where the order of
the arguments supplied to that function was wrong.  The command code
supplied was 0, and the parameter supplied was the command code (and
vice-versa).

Fix this bug.

Signed-off-by: Alex Elder <elder@linaro.org>