soc: qcom: ipa: introduce ipa_reg_init()
Create new functions ipa_reg_init() and ipa_reg_exit(), used to
initialize and shut down the IPA register subsystem.

The new function will initialize the ipahal_ctx->base field, meaning
ipahal_init() doesn't have to.  Remove the argument to ipahal_init()
therefore.

This is the first in a series of commits to separate the register
access functionality from the rest of the IPA HAL code.

Signed-off-by: Alex Elder <elder@linaro.org>