soc: qcom: ipa: kill ipa_alloc_pkt_init()
The whole purpose of ipa_alloc_pkt_init() was to initialize the
entries in the ipa_context->pkt_init_imm[] array.  That array is
gone now, so ipa_alloc_pkt_init() can go away too.  This also means
the pkt_init_mem field of the ipa_context structure unused, so get
rid of that too.

Signed-off-by: Alex Elder <elder@linaro.org>