soc: qcom: ipa: kill IPA_REG_BASE_OFFSET
Currently, all accesses to IPA physical memory add IPA_REG_BASE_OFFSET
in order to translate the value defined by the "ipa-base" property
in the DTB into actual IPA memory range.  The value of that offset
is 0x00040000.

The only spot I can find that would need the address *without*
IPA_REG_BASE_OFFSET added is not present in the IPA code now. that
is the offset advertised by the microcontroller as the "top event
offset," which on my machine currently produces value 0x00047080.

Rather than using the "IPA wrapper" and then adding this offset,
have the range in the DTB define the actual space needed (i.e.,
"pre-offset" by 0x00004000).  At the point we add back the "top
microcontroller event offset" functionality back we can *subtract*
0x00040000 from that address in order to get things into the right
range.

Rename the ipa_wrapper_base field of the IPA context to be ipa_phys
to make it more obvious it no longer represents the same thing.  And
rename the DTB property to be "ipa" rather than "ipa-base" for a
similar reason.  (Also because Rob Herring said that was necessary.)

Signed-off-by: Alex Elder <elder@linaro.org>