soc: qcom: ipa: use PIPE_BITMAP_BMSK
There is a PIPE_BITMAP_BMSK field mask defined for the
IPA_AGGR_FORCE_CLOSE register, but it's not used.

Define a new ipa_reg_aggr_force_close structure to hold the pipe
bitmap value to save to the IPA_AGGR_FORCE_CLOSE register.  Define
ipareg_construct_aggr_force_close() as the construct function for
that register, and use ipahal_write_reg_fields() to write it rather
than just writing the register value directly in the two places it's
written.

Signed-off-by: Alex Elder <elder@linaro.org>