soc: qcom: ipa: more field mask reformatting
Reorder the expressions within construct and parse functions so the
fields operations are performed in the order the field masks are
defined.

Reformat field mask definitions so there's at most one tab following
the longest symbol name in a group.

Align a few wrapped statements with the open parenthesis in the
first line (to satisfy checkpatch.pl).

Signed-off-by: Alex Elder <elder@linaro.org>