soc: qcom: ipa: fix a bug in ipareg_parse_endp_init_aggr_n()
In ipareg_parse_endp_init_aggr_n(), the AGGR_EN field mask is two
bits wide.  When extracting that field value from that register,
the value saved is Boolean, dependent on whether the value in the
field is IPA_ENABLE_AGGR.  Fix this so the actual value in the field
is returned.

Note: it's possible some "full IPA" functionality depends on this
behavior, but for Chrome we do not configure the endpoint for
deaggregation so it does not affect the current code.

Signed-off-by: Alex Elder <elder@linaro.org>