soc: qcom: ipa: arguments are in range in ipahal_hdr_init_local_pyld()
The size of the DMA memory passed to ipahal_hdr_init_local_pyld()
is always less than 2^12, so the check_too_big() call always returns
false.  Replace that call with an assertion instead.  The offet
value passed to ipahal_hdr_init_local_pyld() is always less than
2^16, so the check_too_big() call for that always returns false as
well, so replace that call with an assertion too.

Only hdr_init_local_cmd() calls ipahal_hdr_init_local_pyld() and the
offset and the size of the memory it passes are the same as the
values it receives as arguments.
  - hdr_init_local_cmd() calls ipahal_hdr_init_local_pyld() twice,
    and the arguments are in range:
      offset:
        IPA_MEM_MODEM_HDR_OFST = 0x688 (11 bits required)
        IPA_MEM_APPS_HDR_OFST = 0x7c8 (11 bits required)
      size:
        IPA_MEM_MODEM_HDR_SIZE = 0x140 (9 bits required)
        IPA_MEM_APPS_HDR_SIZE = 0x0 (1 bit required)

Signed-off-by: Alex Elder <elder@linaro.org>