soc: qcom: ipa: dest_pipe_idx is in range in ipahal_ip_packet_init_pyld()
The dest_pipe_idx value passed to ipahal_ip_packet_init_pyld() has
to fit in a 5 bit field.  A check_too_big() call in that function
checks this, but we can simply replace that check with an assertion.

There are currently no callers of ipahal_ip_packet_init_pyld(), so
there's no need to verify this assertion will hold.

Signed-off-by: Alex Elder <elder@linaro.org>