soc: qcom: ipa: don't copy endpoint status info in ipa_cfg_ep_status()
Only ipa_setup_sys_pipe() calls ipa_cfg_ep_status(), and it passes
the address of the status field embedded within an endpoint context
structure.  That status structure is (possibly) populated in
ipa_assign_policy().

Before writing the register, ipa_cfg_ep_status() copies the content
of the endpoint status structure it gets as an argument into the
endpoint's status structure.  This is pointless, because it is
overwriting the source structure with itself.

Just stop updating the endpoint status structure in ipa_cfg_ep_status().

Signed-off-by: Alex Elder <elder@linaro.org>