soc: qcom: ipa: kill ipa_disable_apps_wan_cons_deaggr()
Rather than having ipa_disable_apps_wan_cons_deaggr() quietly verify
that the packet size and count limits supplied are in range, have
the caller do it.  This makes what's going on more obvious.

In addition, setting the generic receive offload flag inside that
function seems a bit hidden.  The rmnet_ipa code should not be
touching the IPA context, but then again, this flag is associated
only with the WAN consumer pipe.  Anyway, set that flag in the
caller, again to make it more obvious what's going on.  We can fix
this flag later.

Signed-off-by: Alex Elder <elder@linaro.org>