soc: qcom: ipa: use a standard macro to compute IPA_GENERIC_RX_BUFF_LIMIT
IPA_GENERIC_RX_BUFF_LIMIT represents the skb space required to hold
standard skb overhead.  It does this by computing the size actually
required to for an 8KB buffer (IPA_RX_BUFFER_SIZE), then subtracts
off the 8KB.

Our buffer is a power-of-2 multiple of pages, so we can instead use
SKB_MAX_ORDER() to determine the maximum size that our buffer will
hold, and deduce the reserved overhead based on that.

Finally, rename this symbol, to be IPA_RX_BUFFER_RESERVED.

Signed-off-by: Alex Elder <elder@linaro.org>