soc: qcom: ipa: redefine IPA_GENERIC_RX_BUFF_SZ()
IPA_GENERIC_RX_BUFF_SZ() returns the amount of space available in a
buffer after standard skb overhead space has been consumed (or
accounted for).  The standard overhead is constant, regardless of
buffer size, so we can use IPA_RX_BUFFER_RESERVED to represent that.

With that, it's much simpler to compute the amount of available
space for a buffer of a given size as (size - IPA_RX_BUFFER_RESERVED).

Reimplement the macro this way.  Rename it IPA_RX_BUFFER_AVAILABLE()
to make it a little more understandable what it represents.

Signed-off-by: Alex Elder <elder@linaro.org>