soc: qcom: ipa: repurpose ipa_aggr_byte_limit_adjust()
Rework ipa_aggr_byte_limit_adjust() so it returns the computed
buffer size, allowing the caller to update the system context buffer
size field.  Then have the caller do the remainder of the
calculation necessary to arrive at the aggregation byte limit value
to use.  Comment the hell out of all of this...

Signed-off-by: Alex Elder <elder@linaro.org>