soc: qcom: ipa: fix an aggregation byte limit bug
The aggregation byte limit is recorded as the a number of kilobytes.
In order to compute the size (in bytes) of a buffer sufficient to
support that aggregation limit we need to convert the recorded value
into bytes.  It's (already) converted back again at the end of
ipa_assign_policy().

Signed-off-by: Alex Elder <elder@linaro.org>