soc: qcom: ipa: drop dst argument from ipa_cfg_ep_mode()
The dst argument to ipa_cfg_ep_mode() is no longer used, so get
rid of it.  That makes the dst argument to ipa_cfg_ep() unused,
so it can go away too.  And that makes the dst argument to
ipa_setup_sys_pipe() unused, so get rid of that as well.

Signed-off-by: Alex Elder <elder@linaro.org>