soc: qcom: ipa: introduce ipa_endp_init_deaggr_prod()
Create new functions to ipa_endp_init_deaggr_prod() to set the
deaggration type for a given endpoint.  And although it only seems
to be used for producer endpoints, create ipa_endp_init_deaggr_cons()
as well.  Use it in the two places that set up producer endpoints.

Add an init_deaggr field to the endpoint context to hold this
information, and use it in ipa_cfg_ep_deaggr().

Finally, rename ipa_cfg_ep_deaggr() to be ipa_endp_init_deaggr_write().

Signed-off-by: Alex Elder <elder@linaro.org>