soc: qcom: ipa: kill struct ipa_ep_cfg
The status field of the ipa_ep_cfg structure is never used.
Removing it removes the last field in that structure.

The only place that structure is uses is for a field in the
ipa_sys_connect_params structure.  But that field is never used, so
it can go away, as can its structure definition.

Signed-off-by: Alex Elder <elder@linaro.org>