soc: qcom: ipa: support new IPA compatible strings
The IPA hardware requires a few registers configured and some
sequencer firmware to be flashed into the hardware during early
initialization.  The ability to do this initialization may be
restricted, and the code executing in the AP non-secure world
might not be permitted to perform this initialization.

Responsibility for this initialization can lie with either the
AP (operating in the Trust Zone secure world), or the modem.
To distinguish between these possible environments we use different
top-level IPA compatible strings in the Device Tree.

This patch puts in basic support for the IPA compatible strings.
Temporarily allow the "old" compatible string "qcom,ipa-sdm845" to
be used, to facilitate testing.

Signed-off-by: Alex Elder <elder@linaro.org>