soc: qcom: ipa: check for IRQ mapping error
If of_irq_get_byname() returns 0, it indicates that there was an
error mapping the IRQ.  This is not a valid result.  So check for
this condition, and have ipa_smp2p_irq() return a negative error
code if it occurs.

Change of_irq_get_byname() to return the (non-zero) IRQ number if
successful, in case the caller wants to record it.

Signed-off-by: Alex Elder <elder@linaro.org>