soc: qcom: ipa: kill IPA_RSRC_GRP_TYPE_SRC_MAX
We no longer need to specify the (outer) number of elements in the
ipa_rsrc_src_grp_config[][] array.  And removing that, we have no
more need for IPA_RSRC_GRP_TYPE_SRC_MAX to be defined.

Similarly we can remove the definition of IPA_RSRC_GRP_TYPE_DST_MAX
if we let the compiler compute the number of elements in the
ipa_rsrc_dst_grp_config[][] array.

Signed-off-by: Alex Elder <elder@linaro.org>