soc: qcom: ipa: introduce enum ipa_rsrc_group
Introduce the ipa_rsrc_group enumerated type, used to define
IPA_RSRC_GROUP_LWA_DL and IPA_RSRC_GROUP_UL_DL.

Also define IPA_RSRC_GROUP_MAX to define the number of valid
enumerated values, and use it in place of IPA_SRC_GROUP_MAX and
IPA_DST_GROUP_MAX (since they are the same).  Note that this reduces
the numeric value used for both of these, but the only effect of
this is to no longer waste space on the initialization arrays.

Signed-off-by: Alex Elder <elder@linaro.org>