soc: qcom: ipa: kill check_too_big()
The last two uses of check_too_big(), in fltrt_init_common(), are
checking information that's determined at runtime, specifically, the
offset of the IPA shared memory area.  This offset must be
representable in 16 bits.

Note first that this would be a hardware design failure if the
condition didn't hold...

In any case, we can verify that for the SDM845, the smem_offset
value returned by hardware is 0x2000 = 8192.  To that we add either:
  IPA_MEM_V4_FLT_HASH_OFST=0x288
  IPA_MEM_V4_RT_NHASH_OFST=0x508
  IPA_MEM_V6_FLT_HASH_OFST=0x388
  IPA_MEM_V6_RT_NHASH_OFST=0x608
The sum in all cases will be less than 0x2800, which is
representable in 14 bits (and therefore in 16 bits).

So get rid of the checks, and get rid of check_too_big().

Signed-off-by: Alex Elder <elder@linaro.org>