soc: qcom: ipa: just allocate memory for firmware
The memory passed to qcom_mdt_load() simply has to be big enough to
hold the firmware that gets loaded.  We need to supply its physical
address because once it's loaded we need to tell the SCM code (TZ)
where to find it.  In this case we have no IOMMU that might require
specifying a particular physical location where the firmware should
be placed.

Signed-off-by: Alex Elder <elder@linaro.org>