soc: qcom: ipa: rename ipa_setup_sys_pipe()
Rename ipa_setup_sys_pipe() to be ipa_ep_setup().  We'll break
things apart more later when we need to differentiate between
"system endpoints" (the ones we support now, which either begin
or end at the AP) and other endpoints (which might be involved in
filtering, or exception processing for other pipes... or something).
Similarly, rename ipa_teardown_sys_pipe() to be ipa_ep_teardown().

Signed-off-by: Alex Elder <elder@linaro.org>