soc: qcom: ipa: kill ipa_get_ep_context()
There is only one caller of ipa_get_ep_context(), used to get access
to the system context pointer associated with an endpoint id in
"ipa_dp.c".  There is another place this information is needed, and
in that case the code basically open-codes what ipa_get_ep_context()
does.

Open-code that function in the one place it's used, and get rid of it.

Signed-off-by: Alex Elder <elder@linaro.org>