soc: qcom: ipa: pass memory struct to ipa_init_rt4()
Pass the address of an ipa_mem structure to ipa_init_rt4(), as well
as ipa_init_rt6().  These functions simply create an empty routing
table that's used as the initial the IPv4 and IPv6 routing tables.
For now, they use a common ipa_mem structure but they otherwise do
the same thing as before.

Signed-off-by: Alex Elder <elder@linaro.org>