soc: qcom: ipa: introduce IPA_MEM_FLT_SIZE
Create new constant IPA_MEM_FLT_SIZE, which represents the size (in
bytes) of a filter table.  It is one more than the number of entries
(an additional first slot is used to hold a bitmap), times the size
of an entry.  All filter tables (used for both IPv4 and IPv6) both
hashed and no-hashed) have the same size, so (for now) use this
value to define all four of those sizes.

Signed-off-by: Alex Elder <elder@linaro.org>