soc: qcom: ipa: rework config_valid()
We now verify in ipa_filter_bitmap_init() that the number of
endpoints that support filtering will not exceed the number of
filter table entries we have.  We therefore know that the filter
bitmap will never have so many set bits that implementing the
filters requires more memory than available in a filter table.
We can therefore get rid of the checks in config_valid() that verify
that at runtime.

Because of that, we no longer need the filter_bitmap argument in
config_valid(), and everything that function checks can be done
at build time.  We can therefore make it return no value, and stop
handling the possibilty it will fail.

Rename that function validate_config(), and move it so it's called
at the very beginning of the probe function.  Move the check that
we're using a 64-bit build into that function.

Signed-off-by: Alex Elder <elder@linaro.org>