soc: qcom: ipa: introduce IPA_MEM_RT_SIZE
Create new constant IPA_MEM_RT_SIZE, which represents the size (in
bytes) of a routing table.  It is the number of entries in a routing
table times the size of an entry.  All routing tables (used for both
IPv4 and IPv6) both hashed and no-hashed) have the same size, so
(for now) use this value to define all four of those sizes.

Signed-off-by: Alex Elder <elder@linaro.org>