soc: qcom: ipa: use IPA_MEM_RT_SIZE
Replace the constants defining the sizes of the IPv4 and IPv6
(hashed and non-hashed) routing tables with IPA_MEM_RT_SIZE.  That
leaves no more use of the four constants, so they can be removed.

It becomes clear that a few sets of assertions are redundant and can
be consolidated.

Signed-off-by: Alex Elder <elder@linaro.org>