soc: qcom: ipa: don't use workqueue for ipa_post_init()
Currently, if the modem is doing early IPA initialization,
ipa_post_init() is called via a workqueue.  I believe this is
because ipa_post_init() may block, and executing it in workqueue
context ensures this won't be a problem.  This is not necessary,
so stop doing it.

When the modem has completed its initialization, it signals an SMP2P
"ipa-post-init" interrupt to the AP.  A threaded interrupt handler
is set up for this IRQ, and that allows the handler function to
block.  So there's no benefit to using the workqueue to ensure the
function can block.

(The only other place ipa_post_init() is called is if Trust Zone is
doing early IPA initialization.  In this case, the call follows the
loading of IPA firmware, which can also block, so we know we must be
running in blockable execution context.)

Signed-off-by: Alex Elder <elder@linaro.org>