soc: qcom: ipa: don't change ipa_send_cmd() return value
In several spots where ipa_send_cmd() is called, if it returns an
error the particular error value is ignored and EFAULT is passed to
the caller instead.  In all of these cases, the return value isn
ignored, so there's no need to translate it.  And it's simpler (and
better) to pass up the actual error code.

Signed-off-by: Alex Elder <elder@linaro.org>