soc: qcom: ipa: verify MDT support at build time
The code supports having either the modem or Trust Zone do
early IPA initialization.  If we use Trust Zone, the MDT firmware
parsing and loading code must be present.  It does no good to check
this at runtime, however.  The config option is selected by the
Kconfig file (and not doing so will produce an eventual build
error), so just get rid of the check.

Signed-off-by: Alex Elder <elder@linaro.org>