soc: qcom: ipa: only clean up post_init_irq if used
In ipa_smp2p_exit(), we are calling ipa_smp2p_irq_exit() to clean up
the post_init_irq unconditionally.  But if we're using Trust Zone
for early initialization, it will not be set.  This is a bug.

Fix this by only calling ipa_smp2p_irq_exit() for the post_init_irq
if its value is non-zero.  Be explicit about assigning its value
regardless of whether Trust Zone or the modem does initialization.

Add a few comments to make it easier to understand a few spots in
the SMP2P interrupt handling code.

Signed-off-by: Alex Elder <elder@linaro.org>