soc: qcom: ipa: don't use namebuf[] in ipa_alloc_workqueue()
There's no need to pre-format the name to use for a workqueue in
ipa_alloc_workqueue(), because alloc_workqueue() already knows how
to do that.  Just pass a format string and the appropriate extra
arguments to alloc_workqueue(), and get rid of the temporary buffer.

This removes the last use of IPA_RESOURCE_NAME_MAX, so get rid of
that definition as well.

Signed-off-by: Alex Elder <elder@linaro.org>