soc: qcom: ipa: change GSI parameterized macro names
Change the names of parameterized register macros so their name
indicates the type of value they use.  Macros that take an EE use
"EE_N" in the name and use "n" as the parameter representing the
EE number.  Macros that take a data channel use "CH_C" in the name
and "c" as the channel id parameter.  Macros that take an event
channel use "EV_CH_E" in the name and "e" to represent the event
channel id.  And finally GSI_INST_RAM_I_OFFS() has parameter "i" to
indicate the instruction index.

Signed-off-by: Alex Elder <elder@linaro.org>