soc: qcom: ipa: have struct ipa_desc use common payload pointer
The payload for an ipa_desc structure is either a void pointer or a
socket buffer fragment.  The two possible values are represented by
a union type.  We can treat the fragment pointer as an opaque
pointer instead, and eliminate the union and the different names.
Use "payload" as the name of either the fragment structure or void
pointer.  This allows a little code in ipa_send() to be simplified.

Do some other rework in ipa_send():
  - Use the descriptor structure rather than the tx_pkt structure
    when mapping the memory for DMA.
  - Store the mapped DMA address in a local variable, and only
    assign it (and other fields) in the tx_pkt structure when we
    it can be filled completely.
  - Don't bother initializing the list links; they're immediately
    passed to list_add_tail(), which overwrites the fields anyway.
  - Use simple "=" assigment rather than "|=" assignment for the
    first values set in a transfer element's flags field.

Signed-off-by: Alex Elder <elder@linaro.org>