soc: qcom: ipa: channel is started in gsi_start_xfer()
The only caller of gsi_start_xfer() is queue_rx_cache().  That
function is only called by ipa_replenish_rx_cache(), which is called
first by ipa_ep_setup() after the channel has been allocated and
started.

It is also called by ipa_rx_common(), which only gets called when a
channel receives a previously-queued packet, meaning the channel is
active.

The last place it's called is ipa_replenish_rx_work_func(), which is
used by ipa_replenish_rx_cache() in the event an error occurrs
attempting to allocate an rx packet wrapper for a queued receive
buffer.  That work on that workqueue is cancelled synchronously in
ipa_ep_teardown() *before* stopping the channel.  So if that work
function gets called, the channel is still active.

So there's no reason to verify the channel is active in
gsi_start_xfer().

Signed-off-by: Alex Elder <elder@linaro.org>