soc: qcom: ipa: don't check wp_local in gsi_start_xfer()
The only caller of gsi_start_xfer() is queue_rx_cache().  The only
time its called is after queueing a transfer element to the
hardware, and furthermore it's called when it's known that at least
IPA_REPL_XFER_THRESH=10 transfer elements have been queued since the
last time the doorbell was rung.  So we know the local write pointer
will be different from the hardware write pointer, and there's no
need to check that condition.

Signed-off-by: Alex Elder <elder@linaro.org>