soc: qcom: ipa: zero channel struct when deallocated
Currently a channel is allocated only once.  But rather than zeroing
the channel structure at allocation time, zero it when it becomes
deallocated, to return it to original state when not in use.  It
will be initially zero, so there's no need to do this on the first
allocation anyway.

Signed-off-by: Alex Elder <elder@linaro.org>