soc: qcom: ipa: pass event ring count to ipa_ep_setup()
Pass the event ring count to ipa_ep_setup() rather than computing it
in that function.  The number of entries in the event ring is always
the same as the number of entries in the channel ring, with the
exception of the APPS_CMD_PROD channel, which has two event ring
entries for each channel ring entry.

Signed-off-by: Alex Elder <elder@linaro.org>