soc: qcom: ipa: kill ipa_gsi_evt_ring_count()
Stop using ipa_gsi_evt_ring_count() to convert a channel entry count
to the number of entries actually used.  It was a mistake to be using
this function for that purpose.

Previously the event ring count supplied was multiplied by two (or
four for the APPS_CMD_PROD pipe).  The multiplication by 2 was an
artifact of an old conversion from BAM to GSI, where BAM ring
entries were 8 bytes and GSI entries are 16.

Instead, have the caller supply the actual number of entries to use
for the channel event ring.  In the process, we are halving the
number of entries used for the APPS_CMD_PROD pipe.  This is fixing a
mistake that went in at some past point that mistakenly doubled the
number used.

The result of this returns the number of entries for both channel
and event rings to what they were in the down stream code:
  APPS_CMD_PROD: 256 channel entries, 512 event entries
  APPS_LAN_CONS: 256 channel entries, 256 event entries
  APPS_WAN_PROD: 512 channel entries, 512 event entries
  APPS_WAN_CONS: 256 channel entries, 256 event entries

Signed-off-by: Alex Elder <elder@linaro.org>