soc: qcom: ipa: kill ipa_sys_connect_param->napi_enabled
The napi_enabled flag of the ipa_sys_connect_param structure is only
true for the APPS_WAN_CONS client.  Rather than passing a separate
flag in the structure, just assign the endpoint's napi_enabled flag
based on the client id passed.

Eventually everything will be NAPI, but for now this simplifes
things a bit.

Signed-off-by: Alex Elder <elder@linaro.org>