soc: qcom: ipa: kill struct ipa_sys_connect_params
There remain only two fields left in the ipa_sys_connect_params
structure.  They are a callback function and a private pointer value
to pass to it.  The structure is used only for passing parameters to
ipa_ep_setup().  Get rid of the structure, and just add two more
arguments that provide the function pointer and void pointer
arguments.

Signed-off-by: Alex Elder <elder@linaro.org>