soc: qcom: ipa: kill gsi_channel_props->low_weight
Rather than passing the scheduling priority weight used for a
channel in as a configuration property, pass in a Boolean flag
indicating whether the channel should get priority.  Currently, only
one channel gets priority (in which case it is assigned maximum
priority); all others get assigned priority 0.  Record the flag
value in the channel structure, not in the properties.

Signed-off-by: Alex Elder <elder@linaro.org>