soc: qcom: ipa: kill gsi_channel_props->user_data
Get rid of the user_data field of teh gsi_channel_props structure.
It is used only for CONS channels, passed as an argument to the
notify callback function.

Pass the value to use to ipa_alloc_channel(), and record this opaque
value in the channel structure instead, using the name notify_data.

Signed-off-by: Alex Elder <elder@linaro.org>