soc: qcom: ipa: stop using use_db_engine channel property
The only thing left in the channel properties structure is the
use_db_engine flag.  That flag is always true initially, and is set
temporarily to false while clearing the pipe during the hardware
workaround for reset while aggreation is open.

Passing a Boolean to gsi_program_channel() indicating whether to
program the channel with the doorbell enabled or not allows the
workaround to be done without the need to preserve that flag
persistently with a channel..

Signed-off-by: Alex Elder <elder@linaro.org>